Added a method for yearly datasets
This is still subject to change due to the seemly risky/bias implementation. But for now it is a method in place to maintain speeding up this project.

Also, updated the dataset with the latest data, only added the 2024 spending. The inforamtion was found on the same website the data was pulled from.